Time: 3 ms (37.77%) | Memory: 40.3 MB (58.68%) - LeetSync

+++ 238-product-of-array-except-self/product-of-array-except-self.cpp
@@ -1,16 +1,16 @@
 class Solution {
 public:
     vector<int> productExceptSelf(vector<int>& nums) {
-        int n = (int)nums.size();
+        int n = nums.size();
         vector<int> ans(n, 1);
-        long long prefix = 1;
+        int prefix = 1;
         for (int i = 0; i < n; i++) {
-            ans[i] = (int)prefix;
+            ans[i] = prefix;
             prefix *= nums[i];
         }
-        long long suffix = 1;
+        int suffix = 1;
         for (int i = n - 1; i >= 0; i--) {
-            ans[i] = (int)( (long long)ans[i] * suffix );
+            ans[i] = (ans[i] * suffix );
             suffix *= nums[i];
         }
